sandbox compiler, improve CE support and add untested OLE support

+++ judges/c++17/info.py
@@ -6,4 +6,5 @@ file_ext = ".cpp"
 
 def garbage_collector():
     os.remove("submission.cpp")
-    os.remove("submission")
+    if "submission" in os.listdir():
+        os.remove("submission")
